backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.0
pydantic==2.10.5
pydantic-settings==2.7.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0  # For mocking utilities
httpx==0.26.0  # For TestClient (updated version)

# AI/ML Libraries
openai==1.59.8
anthropic==0.42.0
langchain==0.3.18
langchain-community==0.3.18
sentence-transformers==3.4.1
langchain-openai==0.3.2
langchain-anthropic==0.3.4
langgraph==0.2.50

# Vector DB
pinecone-client==3.0.2

# SEC Filings & Text Processing
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pypdf==4.0.1
python-multipart==0.0.6

# SEC Filings Downloader
sec-edgar-downloader==5.0.2
